requirements.txt
astroid==2.15.0
attrs==21.2.0
bcrypt==4.0.1
blinker==1.4
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.4
command-not-found==0.3
coverage==7.2.2
cryptography==3.4.8
dbus-python==1.2.18
dill==0.3.6
distro==1.7.0
distro-info==1.1+ubuntu0.1
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
graphviz==0.20.1
greenlet==2.0.1
gunicorn==21.2.0
httplib2==0.20.2
idna==3.4
importlib-metadata==4.6.4
iniconfig==1.1.1
isort==5.12.0
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
joblib==0.17.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy-object-proxy==1.9.0
MarkupSafe==2.1.1
mccabe==0.7.0
more-itertools==8.10.0
netifaces==0.11.0
nltk==3.7
numpy==1.24.0
oauthlib==3.2.0
packaging==21.3
platformdirs==3.1.1
pluggy==0.13.0
psutil==5.9.0
psycopg2==2.9.7
py==1.10.0
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pylint==2.17.0
pyparsing==2.4.7
pytest==6.2.5
python-apt==2.4.0+ubuntu2
python-dotenv==0.21.1
PyYAML==5.4.1
regex==2021.11.10
requests==2.28.2
SecretStorage==3.3.1
simplejson==3.17.6
six==1.16.0
SQLAlchemy==1.4.46
systemd-python==234
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
tqdm==4.57.0
typing_extensions==4.5.0
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.14
wadllib==1.3.6
Werkzeug==2.2.2
wrapt==1.15.0
zipp==1.0.0
